*** empty log message ***
git-svn-id: file:///usr/local/svn/fsc2@3527 934aee4a-8510-0410-ae50-ec5b2ea4bb86

+++ trunk/fsc2/src/util.c
@@ -61,9 +61,9 @@ char *get_string( const char *fmt, ... )
 			continue;
 		}
 
-		if ( wr + 1 > len )        /* newer glibs return the number of chars */
-		{                          /* needed, not counting the trailing '\0' */
-			len = wr + 1;
+		if ( ( size_t ) wr + 1 > len )   /* newer glibs return the number of */
+		{	                             /* chars needed, not counting the   */
+			len = wr + 1;                /* trailing '\0'                    */
 			continue;
 		}
